# add proper type hints for performance