Added proper type hints for the compiler, doubling performance w/o IO, and roughly 1.5x performance with io