@leetwinski 2024-01-17
Added proper type hints for compiler, doubling performance w/o IO, and roughly 1.5x performance with io