@jinyus 2024-01-17
I can't get this to work on linux/docker so I can't post the comparison.