@jinyus 2024-01-17
This is faster but both  throws a stack overflow error on the 60k round:

old:
```
Common Lisp (SBCL):

        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107008k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.35s memory: 107008k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107264k
        Processing time (w/o IO): 0.26 s
        Processing time total: 0.35 s
        total: 0.35s memory: 107008k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107136k
        Processing time (w/o IO): 0.25 s
        Processing time total: 0.34 s
        total: 0.35s memory: 107008k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107136k
        Processing time (w/o IO): 0.25 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107136k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107136k
        Processing time (w/o IO): 0.24 s
        Processing time total: 0.34 s
        total: 0.34s memory: 107136k

Common Lisp (SBCL):

        Processing time (w/o IO): 2.34 s
        Processing time total: 2.69 s
        total: 2.70s memory: 116408k
        Processing time (w/o IO): 2.35 s
        Processing time total: 2.70 s
        total: 2.71s memory: 116528k

stackoverflow on 3rd run.
```

This PR:
```
Common Lisp (SBCL):

        Processing time (w/o IO): 0.16 s
        Processing time total: 0.25 s
        total: 0.26s memory: 107136k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 106880k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 107136k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 107008k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.27 s
        total: 0.28s memory: 107136k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.25 s
        total: 0.26s memory: 107008k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 107136k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.25 s
        total: 0.26s memory: 107136k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 107008k
        Processing time (w/o IO): 0.16 s
        Processing time total: 0.26 s
        total: 0.26s memory: 107008k

Common Lisp (SBCL):

        Processing time (w/o IO): 0.99 s
        Processing time total: 1.34 s
        total: 1.35s memory: 116664k
        Processing time (w/o IO): 1.00 s
        Processing time total: 1.54 s
        total: 1.55s memory: 116540k

Common Lisp (SBCL):

        INFO: Control stack guard page unprotected
        Control stack guard page temporarily disabled: proceed with caution
        Unhandled SB-KERNEL::CONTROL-STACK-EXHAUSTED in thread #<SB-THREAD:THREAD tid=108 "main thread" RUNNING
                                                                  {10019F8003}>:
          Control stack exhausted (no more space for function call frames).
        This is probably due to heavily nested or infinitely recursive function
        calls, or a tail call that SBCL cannot or has not optimized away.

        PROCEED WITH CAUTION.

        Backtrace for: #<SB-THREAD:THREAD tid=108 "main thread" RUNNING {10019F8003}>
        0: (SB-DEBUG::DEBUGGER-DISABLED-HOOK #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}> #<unused argument> :QUIT T)
        1: (SB-DEBUG::RUN-HOOK *INVOKE-DEBUGGER-HOOK* #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}>)
        2: (INVOKE-DEBUGGER #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}>)
        3: (ERROR SB-KERNEL::CONTROL-STACK-EXHAUSTED)
        4: (SB-KERNEL::CONTROL-STACK-EXHAUSTED-ERROR)
        5: ("foreign function: call_into_lisp_")
        6: ("foreign function: post_signal_tramp")
        7: (SB-IMPL::GROW-HASH-TABLE 0)
        8: ("foreign function: #x0")
        9: ((FLET SB-IMPL::INSERT-AT :IN "SYS:SRC;CODE;TARGET-HASH-TABLE.LISP") #<unavailable argument> #<HASH-TABLE :TEST EQUAL :COUNT 0 {10031FCD73}> "_id" 1479427920 NIL "9wz9htbt")
        10: (SB-IMPL::PUTHASH/EQUAL #<unavailable argument> #<unavailable argument> #<unavailable argument>)
        11: (SB-KERNEL:%PUTHASH "_id" #<HASH-TABLE :TEST EQUAL :COUNT 0 {10031FCD73}> "9wz9htbt")
        12: (COM.INUOE.JZON::%PARSE #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D99AB}> #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D9E3B}> #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D99CB}> #<FUNCTION (LAMBDA (COM.INUOE.JZON::KEY) :IN COM.INUOE.JZON::%MAKE-STRING-POOL) {10019D8BEB}> 128 NIL NIL NIL)
        13: (COM.INUOE.JZON:PARSE #<unavailable argument> :MAX-DEPTH #<unavailable argument> :ALLOW-COMMENTS #<unavailable argument> :ALLOW-TRAILING-COMMA #<unavailable argument> :ALLOW-MULTIPLE-CONTENT #<unavailable argument> :MAX-STRING-LENGTH #<unavailable argument> :KEY-FN #<unavailable argument>)
        14: (MAIN)
        15: ((FLET SB-UNIX::BODY :IN SB-IMPL::START-LISP))
        16: ((FLET "WITHOUT-INTERRUPTS-BODY-3" :IN SB-IMPL::START-LISP))
        17: (SB-IMPL::%START-LISP)

        unhandled condition in --disable-debugger mode, quitting
        Command exited with non-zero status 1
        total: 0.07s memory: 51712k
        INFO: Control stack guard page unprotected
        Control stack guard page temporarily disabled: proceed with caution
        Unhandled SB-KERNEL::CONTROL-STACK-EXHAUSTED in thread #<SB-THREAD:THREAD tid=113 "main thread" RUNNING
                                                                  {10019F8003}>:
          Control stack exhausted (no more space for function call frames).
        This is probably due to heavily nested or infinitely recursive function
        calls, or a tail call that SBCL cannot or has not optimized away.

        PROCEED WITH CAUTION.

        Backtrace for: #<SB-THREAD:THREAD tid=113 "main thread" RUNNING {10019F8003}>
        0: (SB-DEBUG::DEBUGGER-DISABLED-HOOK #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}> #<unused argument> :QUIT T)
        1: (SB-DEBUG::RUN-HOOK *INVOKE-DEBUGGER-HOOK* #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}>)
        2: (INVOKE-DEBUGGER #<SB-KERNEL::CONTROL-STACK-EXHAUSTED {10031FCEB3}>)
        3: (ERROR SB-KERNEL::CONTROL-STACK-EXHAUSTED)
        4: (SB-KERNEL::CONTROL-STACK-EXHAUSTED-ERROR)
        5: ("foreign function: call_into_lisp_")
        6: ("foreign function: post_signal_tramp")
        7: (SB-IMPL::GROW-HASH-TABLE 0)
        8: ("foreign function: #x0")
        9: ((FLET SB-IMPL::INSERT-AT :IN "SYS:SRC;CODE;TARGET-HASH-TABLE.LISP") #<unavailable argument> #<HASH-TABLE :TEST EQUAL :COUNT 0 {10031FCD73}> "_id" 1479427920 NIL "9wz9htbt")
        10: (SB-IMPL::PUTHASH/EQUAL #<unavailable argument> #<unavailable argument> #<unavailable argument>)
        11: (SB-KERNEL:%PUTHASH "_id" #<HASH-TABLE :TEST EQUAL :COUNT 0 {10031FCD73}> "9wz9htbt")
        12: (COM.INUOE.JZON::%PARSE #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D99AB}> #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D9E3B}> #<FUNCTION (LAMBDA NIL :IN COM.INUOE.JZON::%MAKE-FNS-STREAM) {10019D99CB}> #<FUNCTION (LAMBDA (COM.INUOE.JZON::KEY) :IN COM.INUOE.JZON::%MAKE-STRING-POOL) {10019D8BEB}> 128 NIL NIL NIL)
        13: (COM.INUOE.JZON:PARSE #<unavailable argument> :MAX-DEPTH #<unavailable argument> :ALLOW-COMMENTS #<unavailable argument> :ALLOW-TRAILING-COMMA #<unavailable argument> :ALLOW-MULTIPLE-CONTENT #<unavailable argument> :MAX-STRING-LENGTH #<unavailable argument> :KEY-FN #<unavailable argument>)
        14: (MAIN)
        15: ((FLET SB-UNIX::BODY :IN SB-IMPL::START-LISP))
        16: ((FLET "WITHOUT-INTERRUPTS-BODY-3" :IN SB-IMPL::START-LISP))
        17: (SB-IMPL::%START-LISP)

        unhandled condition in --disable-debugger mode, quitting
        Command exited with non-zero status 1
        total: 0.07s memory: 51840k
```